+++ tests/Unit/Scenarios/ImportPdfViaFile.php
@@ -19,6 +19,7 @@ class ImportPdfViaFile extends externalFilesTests {
 	 * @return void
 	 */
 	public function test_import_pdf_via_file(): void {
+		add_filter( 'efml_file_base', array( $this, 'set_plugin_path_as_base' ) );
 		$result = \ExternalFilesInMediaLibrary\ExternalFiles\Import::get_instance()->add_url( self::get_test_file( 'pdf', 'file' ) );
 		$this->assertIsBool( $result );
 		$this->assertTrue( $result );